added boomsheets animation file parser and it works
