+++ BizHawk-2.5.2/canodumb.animation
@@ -1,0 +1,44 @@
+VERSION="2.0"
+
+# File "canodumb.animation" generated by BoomSheets Atlas 06/30/2018 @ 07:06 PM
+
+imagePath="canodumb_atlas.png"
+
+animation state="IDLE_1"
+frame duration="0.1" x="31" y="2" w="39" h="55" originx="19" originy="49"
+
+animation state="SHOOT_1"
+frame duration="0.1" x="31" y="2" w="39" h="55" originx="19" originy="49"
+frame duration="0.1" x="74" y="2" w="39" h="55" originx="19" originy="49"
+frame duration="0.1" x="117" y="2" w="41" h="55" originx="19" originy="49"
+frame duration="0.1" x="162" y="2" w="41" h="55" originx="19" originy="49"
+
+animation state="IDLE_2"
+frame duration="0.1" x="31" y="57" w="39" h="55" originx="19" originy="49"
+
+animation state="SHOOT_2"
+frame duration="0.1" x="31" y="57" w="39" h="55" originx="19" originy="49"
+frame duration="0.1" x="74" y="57" w="39" h="55" originx="19" originy="49"
+frame duration="0.1" x="117" y="57" w="41" h="55" originx="19" originy="49"
+frame duration="0.1" x="162" y="57" w="41" h="55" originx="19" originy="49"
+
+animation state="IDLE_3"
+frame duration="0.1" x="31" y="112" w="39" h="55" originx="19" originy="49"
+
+animation state="SHOOT_3"
+frame duration="0.1" x="31" y="112" w="39" h="55" originx="19" originy="49"
+frame duration="0.1" x="74" y="112" w="39" h="55" originx="19" originy="49"
+frame duration="0.1" x="117" y="112" w="41" h="55" originx="19" originy="49"
+frame duration="0.1" x="162" y="112" w="41" h="55" originx="19" originy="49"
+
+animation state="CURSOR"
+frame duration="0.1" x="5" y="185" w="37" h="39" originx="19" originy="35"
+
+animation state="SMOKE"
+frame duration="0.03" x="51" y="188" w="19" h="29" originx="14" originy="15"
+frame duration="0.03" x="77" y="184" w="22" h="41" originx="17" originy="20"
+frame duration="0.03" x="106" y="181" w="26" h="47" originx="17" originy="23"
+frame duration="0.03" x="139" y="173" w="27" h="65" originx="17" originy="32"
+frame duration="0.03" x="170" y="171" w="25" h="67" originx="20" originy="34"
+frame duration="0.03" x="206" y="169" w="25" h="67" originx="20" originy="34"
+
