+++ BizHawk-2.5.2/config.ini
@@ -1631,7 +1631,7 @@
       }
     ],
     "UseRecentForRoms": false,
-    "LastRomPath": "C:\\Users\\corba\\Documents\\GitHub\\BBN3-netcode\\BizHawk-2.5.2\\BBN3",
+    "LastRomPath": "X:\\Big_Code\\BBN3-netcode\\BizHawk-2.5.2\\BBN3",
     "FirmwaresPathFragment": "./Firmware",
     "TempFilesFragment": ""
   },
@@ -1666,8 +1666,8 @@
   "SaveWindowPosition": true,
   "StartPaused": false,
   "StartFullscreen": false,
-  "MainWndx": 187,
-  "MainWndy": 338,
+  "MainWndx": 169,
+  "MainWndy": 239,
   "RunInBackground": true,
   "AcceptBackgroundInput": false,
   "AcceptBackgroundInputControllerOnly": false,
@@ -1818,7 +1818,7 @@
   "DispCropRight": 0,
   "DispCropBottom": 0,
   "SoundOutputMethod": 0,
-  "SoundEnabled": true,
+  "SoundEnabled": false,
   "SoundEnabledNormal": true,
   "SoundEnabledRWFF": true,
   "MuteFrameAdvance": true,
