+++ BizHawk-2.5.2/config2.ini
@@ -1631,7 +1631,7 @@
       }
     ],
     "UseRecentForRoms": false,
-    "LastRomPath": "C:\\Users\\corba\\Documents\\GitHub\\BBN3-netcode\\BizHawk-2.5.2\\BBN3",
+    "LastRomPath": "X:\\Big_Code\\BBN3-netcode\\BizHawk-2.5.2\\BBN3",
     "FirmwaresPathFragment": "./Firmware",
     "TempFilesFragment": ""
   },
@@ -1667,8 +1667,8 @@
   "SaveWindowPosition": true,
   "StartPaused": false,
   "StartFullscreen": false,
-  "MainWndx": 701,
-  "MainWndy": 337,
+  "MainWndx": 766,
+  "MainWndy": 255,
   "RunInBackground": true,
   "AcceptBackgroundInput": true,
   "AcceptBackgroundInputControllerOnly": false,
@@ -1899,7 +1899,7 @@
     "BizHawk.Client.EmuHawk.LuaConsole": {
       "_wndx": 700,
       "_wndy": 715,
-      "Width": 497,
+      "Width": 495,
       "Height": 327,
       "SaveWindowPosition": true,
       "TopMost": false,
@@ -2007,7 +2007,7 @@
         },
         "ReloadOnScriptFileChange": false,
         "ToggleAllIfNoneSelected": true,
-        "SplitDistance": 114,
+        "SplitDistance": 113,
         "DisableLuaScriptsOnLoad": false
       }
     },
